refactoring specs, object reference and other minor improvements
